+++ util/jsonrpc-types/CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0](https://github.com/nervosnetwork/ckb/compare/ckb-jsonrpc-types-v1.1.2...ckb-jsonrpc-types-v1.2.0) - 2026-07-28
+
+### Added
+
+- *(rpc)* include block hash in live cell response (#5269) (by @quake)
+
+### Contributors
+
+- @quake
+
 ## [1.1.2](https://github.com/nervosnetwork/ckb/compare/ckb-jsonrpc-types-v1.1.1...ckb-jsonrpc-types-v1.1.2) - 2026-06-08
 
 ### Changed