+++ util/logger-service/CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.2](https://github.com/nervosnetwork/ckb/compare/ckb-logger-service-v1.1.1...ckb-logger-service-v1.1.2) - 2026-07-28
+
+### Fixed
+
+- Fix/unbounded memory growth (#5292) (by @chenyukang)
+
+### Contributors
+
+- @chenyukang
+
 ## [1.1.1](https://github.com/nervosnetwork/ckb/compare/ckb-logger-service-v1.1.0...ckb-logger-service-v1.1.1) - 2026-06-08
 
 ### Changed